Stricter, Faster Dup Matching
With more decompiled code we have more of an idea what is duplicated in
each overlay. Currently dups are only considered based on instruction
opcode, which is relatively coarse and produces false positives.
The other useful feature for local development - finding similar
functions with a lower match - threshold becomes impractical as more
functions are evaluated and runtimes grow.

This changes addresses a few things:

* instructions are masked based on format which produces more accurate
  matches
* threshold has been increased to 98% for reporting to filter out "near"
  neighbors which are often different
* function size is considered before and as an approximation to
  Levenshtein distance to avoid the cost of computing similarity
  which would never reach the requested threshold

The increased data from instructions addresses situations like
`EntityRoomForeground` and `EntityBackgroundBlock` or `EntityBloodDrips`
and `EntityOlroxDrool` matching. These are now different enough that
even with a 90% similarity match these fall into separate groups.

The higher match prevents low quality matches which are grouped early
from preventing closer matches later on. I was optimizing for the
`make-config` case, but this has value in the dups report as well.
Threshold can be reduced further to find "close" neighbors.

Performing a size comparison before Levenshstein results in a
significant reduction in runtime. Previously `dups` took around 7
minutes on my local machine. It now runs in around 23 seconds. Since any
delta in size would produce the equivalent number of "edits" when
computing Levenshtein, this produces a fast approximation which
eliminates the majority of similarity calculations. Because there are
no cases where a calculated Levenshtein can be smaller than the
difference in character count, this is a free speedup without any effect
to the output result.

There was also a grouping bug where a function would be put into the
least matching group that was higher than the threshold. This is harder
to identify how it affected the output, but similar clusters should
group more closely.

+++ tools/dups/src/types.rs
@@ -2,7 +2,7 @@
 pub struct Function {
     pub name: String,
     pub ops: Vec<Instruction>,
-    pub key: Vec<u8>,
+    pub key: Vec<u32>,
     pub dir: String,
     pub file: String,
     pub similarity: f64,
